Added euclidean and manhattan metric

+++ Kod/src/main/java/tech/szymanskazdrzalik/ksr/ksr1/metric/ChebyshevMetric.java
@@ -2,7 +2,7 @@ package tech.szymanskazdrzalik.ksr.ksr1.metric;
 
 public class ChebyshevMetric implements Metric {
     @Override
-    public float calculateDistance(double[] article, double[] trainingArticle) {
+    public double calculateDistance(double[] article, double[] trainingArticle) {
         return 0;
     }
 }
